tarce: add support for trace
Enable tracing capability for debugging

+++ demos/sd3403/build/build_fetch.sh
@@ -1,19 +1,23 @@
 pushd ../../../src/net/
 echo "################# get soem #################"
-rm -rf ./soem
-git clone --depth=1 --filter=blob:none --sparse -b revert-2058de2 https://atomgit.com/openeuler/oee_archive.git
-cd oee_archive
-git sparse-checkout init --cone
-git sparse-checkout set "soem"
-cp soem/SOEM-1.4.0.tar.gz ../
-cd ../
-rm -rf oee_archive
-tar -zxf SOEM-1.4.0.tar.gz
-rm SOEM-1.4.0.tar.gz
-mv SOEM-1.4.0 soem
-cp UniProton-patch-for-soem.patch ./soem
-cd ./soem
-patch -p1 -d . < UniProton-patch-for-soem.patch
-rm -rf ../../../testsuites/soemTest/test
-cp -r test ../../../testsuites/soemTest/
+if [ -d "./soem" ] && [ -f "./soem/CMakeLists.txt" ]; then
+    echo "soem already exists, skip download and patch."
+else
+    rm -rf ./soem
+    git clone --depth=1 --filter=blob:none --sparse -b revert-2058de2 https://atomgit.com/openeuler/oee_archive.git
+    cd oee_archive
+    git sparse-checkout init --cone
+    git sparse-checkout set "soem"
+    cp soem/SOEM-1.4.0.tar.gz ../
+    cd ../
+    rm -rf oee_archive
+    tar -zxf SOEM-1.4.0.tar.gz
+    rm SOEM-1.4.0.tar.gz
+    mv SOEM-1.4.0 soem
+    cp UniProton-patch-for-soem.patch ./soem
+    cd ./soem
+    patch -p1 -d . < UniProton-patch-for-soem.patch
+    rm -rf ../../../testsuites/soemTest/test
+    cp -r test ../../../testsuites/soemTest/
+fi
 popd
