+++ demos/sd3403/build/build_openamp.sh
@@ -1,27 +1,35 @@
 echo "################# git clone libmetal #################"
 pushd ../component
-rm -rf ./libmetal*
-git clone https://gitee.com/src-openeuler/libmetal.git
-mv ./libmetal/libmetal-2022.10.0.tar.gz .
-rm -rf ./libmetal
-tar -zxvf libmetal-2022.10.0.tar.gz
-mv ./libmetal-2022.10.0 ./libmetal
-cp UniProton-patch-for-libmetal.patch ./libmetal
-cd libmetal
-patch -p1 -d . < UniProton-patch-for-libmetal.patch
+if [ -d "./libmetal" ] && [ -f "./libmetal/CMakeLists.txt" ]; then
+    echo "libmetal source already exists, skip download."
+else
+    rm -rf ./libmetal*
+    git clone https://gitee.com/src-openeuler/libmetal.git
+    mv ./libmetal/libmetal-2022.10.0.tar.gz .
+    rm -rf ./libmetal
+    tar -zxvf libmetal-2022.10.0.tar.gz
+    mv ./libmetal-2022.10.0 ./libmetal
+    cp UniProton-patch-for-libmetal.patch ./libmetal
+    cd libmetal
+    patch -p1 -d . < UniProton-patch-for-libmetal.patch
+fi
 popd
 
 echo "################# git clone openamp #################"
 pushd ../component
-rm -rf ./open-amp*
-git clone https://gitee.com/src-openeuler/OpenAMP.git
-mv ./OpenAMP/openamp-2022.10.1.tar.gz .
-rm -rf ./OpenAMP
-tar -zxvf openamp-2022.10.1.tar.gz
-mv ./openamp-2022.10.1 ./open-amp
-cp UniProton-patch-for-openamp.patch ./open-amp
-cd open-amp
-patch -p1 -d . < UniProton-patch-for-openamp.patch
+if [ -d "./open-amp" ] && [ -f "./open-amp/CMakeLists.txt" ]; then
+    echo "open-amp source already exists, skip download."
+else
+    rm -rf ./open-amp*
+    git clone https://gitee.com/src-openeuler/OpenAMP.git
+    mv ./OpenAMP/openamp-2022.10.1.tar.gz .
+    rm -rf ./OpenAMP
+    tar -zxvf openamp-2022.10.1.tar.gz
+    mv ./openamp-2022.10.1 ./open-amp
+    cp UniProton-patch-for-openamp.patch ./open-amp
+    cd open-amp
+    patch -p1 -d . < UniProton-patch-for-openamp.patch
+fi
 popd
 
 echo "######################### build metal #########################"
